Notify V8 when we have many WebGL contexts
This fix is to deal with the WebGL 1.0.2
context-creation-and-destruction test which creates 500 contexts quickly
and relies on JS cleaning up contexts as they go out of scope. WebGL
context objects consume almost no memory on the JS heap, and V8 does not
know what kind of external resources are associated with them. The
garbage collector is likely not running frequently enough to dispose of
the contexts. Calling lowMemoryNotification when we have many contexts
should force a major GC and free up the WebGL resources.

Change-Id: I3c8f51ddeede899e2b6a3d14a87b8c80883dd498
CRs-fixed: 427617